Address review comments and add metrics

+++ runner/runner.go
@@ -4,18 +4,22 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"math"
 	"os"
 	"os/exec"
 	"os/user"
 	"path/filepath"
+	"strings"
 	"sync"
 	"sync/atomic"
 	"syscall"
 	"time"
 
-	"github.com/temporalio/omes/scenarios"
+	"github.com/temporalio/omes/scenario"
 	"github.com/temporalio/omes/shared"
+	"go.opentelemetry.io/otel/attribute"
+	metrics "go.opentelemetry.io/otel/metric/global"
+	"go.opentelemetry.io/otel/metric/instrument"
+	"go.opentelemetry.io/otel/metric/instrument/syncint64"
 	"go.temporal.io/api/batch/v1"
 	"go.temporal.io/api/enums/v1"
 	"go.temporal.io/api/workflowservice/v1"
@@ -25,25 +29,28 @@ import (
 
 const DEFAULT_CONCURRENCY = 10
 
+var Meter = metrics.Meter("omes")
+
 // Options for creating a Runner
 type Options struct {
 	ClientOptions client.Options
 	// ID used for prefixing workflow IDs and determining the task queue
 	RunID    string
-	Scenario scenarios.Scenario
+	Scenario *scenario.Scenario
 }
 
-type RunOptions struct {
-	// Whether to abort immediately after first error
-	FailFast bool
+type RunnerMetrics struct {
+	executeHistogram  syncint64.Histogram
+	scenarioAttribute attribute.KeyValue
 }
 
 type Runner struct {
-	iterationCounter atomic.Uint32
+	iterationCounter atomic.Uint64
 	done             sync.WaitGroup
 	options          Options
 	errors           chan error
 	logger           *zap.SugaredLogger
+	metrics          *RunnerMetrics
 }
 
 // NewRunner instantiates a Runner
@@ -57,21 +64,35 @@ func NewRunner(options Options, logger *zap.SugaredLogger) (*Runner, error) {
 		return nil, errors.New("invalid scenario: iterations and duration are mutually exclusive")
 	}
 
+	executeHistogram, err := Meter.SyncInt64().Histogram(
+		"execute.histogram",
+		instrument.WithUnit("microseconds"),
+		instrument.WithDescription("Excute method histogram"),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create a histogram: %w", err)
+	}
+	scenarioAttribute := attribute.KeyValue{Key: "scenario", Value: attribute.StringValue(options.Scenario.Name)}
+
 	return &Runner{
 		options: options,
 		errors:  make(chan error),
 		logger:  logger,
+		metrics: &RunnerMetrics{
+			executeHistogram:  executeHistogram,
+			scenarioAttribute: scenarioAttribute,
+		},
 	}, nil
 }
 
-func calcConcurrency(iterations uint32, scenario scenarios.Scenario) int {
-	concurrency := int(scenario.Concurrency)
+func calcConcurrency(iterations int, scenario *scenario.Scenario) int {
+	concurrency := scenario.Concurrency
 	if concurrency == 0 {
 		concurrency = DEFAULT_CONCURRENCY
 	}
-	if iterations > 0 {
+	if iterations > 0 && concurrency > iterations {
 		// Don't spin up more coroutines than the number of total iterations
-		concurrency = int(math.Min(float64(concurrency), float64(iterations)))
+		concurrency = iterations
 	}
 	return concurrency
 }
@@ -79,7 +100,7 @@ func calcConcurrency(iterations uint32, scenario scenarios.Scenario) int {
 // Run a scenario.
 // Spins up coroutines according to the scenario configuration.
 // Each coroutine runs the scenario Execute method in a loop until the scenario duration or max iterations is reached.
-func (r *Runner) Run(ctx context.Context, options RunOptions) error {
+func (r *Runner) Run(ctx context.Context) error {
 	c, err := shared.Connect(r.options.ClientOptions, r.logger)
 	if err != nil {
 		return err
@@ -94,11 +115,11 @@ func (r *Runner) Run(ctx context.Context, options RunOptions) error {
 	if duration > 0 {
 		ctx, cancel = context.WithTimeout(ctx, duration)
 	}
+	defer cancel()
 
 	r.done.Add(concurrency)
 
 	startTime := time.Now()
-	hadErrors := false
 	waitChan := make(chan struct{})
 	go func() {
 		r.done.Wait()
@@ -107,23 +128,21 @@ func (r *Runner) Run(ctx context.Context, options RunOptions) error {
 
 	for i := 0; i < concurrency; i++ {
 		logger := r.logger.With("coroID", i)
-		go r.runOne(ctx, logger, c, options)
+		go r.runOne(ctx, logger, c)
 	}
+
+	var accumulatedErrors []string
 
 	for {
 		select {
-		case <-r.errors:
-			if options.FailFast {
-				cancel()
-			}
-			hadErrors = true
-		case <-waitChan:
-			// Cancel to avoid potential context leak
+		case err := <-r.errors:
 			cancel()
-			if hadErrors {
-				return fmt.Errorf("run finished with errors after %s", time.Since(startTime))
+			accumulatedErrors = append(accumulatedErrors, err.Error())
+		case <-waitChan:
+			if len(accumulatedErrors) > 0 {
+				return fmt.Errorf("run finished with errors after %s, errors:\n%s", time.Since(startTime), strings.Join(accumulatedErrors, "\n"))
 			}
-			r.logger.Info("Run complete in ", time.Since(startTime))
+			r.logger.Infof("Run complete in %v", time.Since(startTime))
 			return nil
 		}
 	}
@@ -131,39 +150,38 @@ func (r *Runner) Run(ctx context.Context, options RunOptions) error {
 
 // runOne - where "one" is a single routine out of N concurrent defined for the scenario.
 // This method will loop until context is cancelled or the number of iterations for the scenario have exhuasted.
-func (r *Runner) runOne(ctx context.Context, logger *zap.SugaredLogger, c client.Client, options RunOptions) {
+func (r *Runner) runOne(ctx context.Context, logger *zap.SugaredLogger, c client.Client) {
 	iterations := r.options.Scenario.Iterations
-loop:
+	defer r.done.Done()
 	for {
-		select {
-		case <-ctx.Done():
-			break loop
-		default:
+		if ctx.Err() != nil {
+			return
 		}
-		iteration := r.iterationCounter.Add(1)
+		iteration := int(r.iterationCounter.Add(1))
 		// If the scenario is limited in number of iterations, do not exceed that number
 		if iterations > 0 && iteration > iterations {
 			break
 		}
 		logger.Debugf("Running iteration %d", iteration)
-		run := scenarios.Run{
+		run := scenario.Run{
 			Client:          c,
-			Scenario:        &r.options.Scenario,
+			Scenario:        r.options.Scenario,
 			IterationInTest: iteration,
-			Logger:          logger,
+			Logger:          logger.With("iteration", iteration),
 			ID:              r.options.RunID,
 		}
+
+		startTime := time.Now()
 		if err := r.options.Scenario.Execute(ctx, &run); err != nil {
-			err = fmt.Errorf("iteration %d failed: %v", iteration, err)
+			duration := time.Since(startTime)
+			r.metrics.executeHistogram.Record(ctx, duration.Microseconds(), r.metrics.scenarioAttribute)
+			err = fmt.Errorf("iteration %d failed: %w", iteration, err)
 			logger.Error(err)
 			r.errors <- err
 			// Even though context will be cancelled by the runner, we break here to avoid needlessly running another iteration
-			if options.FailFast {
-				break
-			}
+			break
 		}
 	}
-	r.done.Done()
 }
 
 type CleanupOptions struct {
@@ -192,7 +210,7 @@ func (r *Runner) Cleanup(ctx context.Context, options CleanupOptions) error {
 		return err
 	}
 	defer c.Close()
-	taskQueue := scenarios.TaskQueueForRunID(&r.options.Scenario, r.options.RunID)
+	taskQueue := r.options.Scenario.TaskQueueForRunID(r.options.RunID)
 	jobId := taskQueue
 	// Clean based on task queue to avoid relying on search attributes and reducing the requirements of this framework.
 	// Not escaping the value here and living with the consequences.
@@ -225,9 +243,16 @@ func (r *Runner) Cleanup(ctx context.Context, options CleanupOptions) error {
 		case enums.BATCH_OPERATION_STATE_COMPLETED:
 			return nil
 		case enums.BATCH_OPERATION_STATE_RUNNING:
-			time.Sleep(options.PollInterval)
+			select {
+			case <-time.After(options.PollInterval):
+				// go to next loop iteration
+			case <-ctx.Done():
+				return ctx.Err()
+			}
 		case enums.BATCH_OPERATION_STATE_UNSPECIFIED:
 			return fmt.Errorf("invalid batch state: %s - reason: %s", response.State, response.Reason)
+		default:
+			return fmt.Errorf("unexepcted batch state: %s - reason: %s", response.State, response.Reason)
 		}
 	}
 }
@@ -260,18 +285,31 @@ type WorkerOptions struct {
 	Language    string
 	TLSCertPath string
 	TLSKeyPath  string
+	// Time to wait before killing the worker process after sending SIGTERM in case it doesn't gracefully shut down.
+	// Default is 30 seconds.
+	GracefulShutdownDuration time.Duration
+	//
+	RetainBuildDir bool
 }
 
-// StartWorker prepares (e.g. builds) and starts a worker for a given language.
+// RunWorker prepares (e.g. builds) and run a worker for a given language.
 // The worker process will be killed with SIGTERM when the given context is cancelled.
-func (r *Runner) StartWorker(ctx context.Context, options WorkerOptions) error {
+// If the worker process does not exit after options.GracefulShutdownDuration, it will get a SIGKILL
+func (r *Runner) RunWorker(ctx context.Context, options WorkerOptions) error {
 	var args []string
 	tmpDir, err := os.MkdirTemp(os.TempDir(), "omes-build-")
-	defer os.RemoveAll(tmpDir)
-
 	if err != nil {
-		return fmt.Errorf("failed to create temp dir: %v", err)
+		return fmt.Errorf("failed to create temp dir: %w", err)
 	}
+	r.logger.Infof("Created worker build dir %s", tmpDir)
+	if options.RetainBuildDir {
+		defer os.RemoveAll(tmpDir)
+	}
+	gracefulShutdownDuration := options.GracefulShutdownDuration
+	if gracefulShutdownDuration == 0 {
+		gracefulShutdownDuration = 30 * time.Second
+	}
+
 	switch options.Language {
 	case "go":
 		outputPath := filepath.Join(tmpDir, "worker")
@@ -282,7 +320,7 @@ func (r *Runner) StartWorker(ctx context.Context, options WorkerOptions) error {
 			outputPath,
 			"--server-address", r.options.ClientOptions.HostPort,
 			"--namespace", r.options.ClientOptions.Namespace,
-			"--task-queue", scenarios.TaskQueueForRunID(&r.options.Scenario, r.options.RunID),
+			"--task-queue", r.options.Scenario.TaskQueueForRunID(r.options.RunID),
 		}
 		if options.TLSCertPath != "" && options.TLSKeyPath != "" {
 			args = append(args, "--tls-cert-path", options.TLSCertPath, "--tls-key-path", options.TLSKeyPath)
@@ -291,7 +329,8 @@ func (r *Runner) StartWorker(ctx context.Context, options WorkerOptions) error {
 		return fmt.Errorf("language not supported: '%s'", options.Language)
 	}
 
-	runErrorChan := make(chan error)
+	runErrorChan := make(chan error, 1)
+	// Inentionally not using CommandContext since we want to kill the worker gracefully (using SIGTERM).
 	cmd := exec.Command(args[0], args[1:]...)
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
@@ -310,9 +349,17 @@ func (r *Runner) StartWorker(ctx context.Context, options WorkerOptions) error {
 		r.logger.Infof("Sending SIGTERM to worker, PID: %d", cmd.Process.Pid)
 		err := cmd.Process.Signal(syscall.SIGTERM)
 		if err != nil {
-			r.logger.Fatalf("failed to kill worker: %v", err)
+			r.logger.Fatalf("failed to kill worker: %w", err)
 		}
-		return <-runErrorChan
+		select {
+		case err := <-runErrorChan:
+			return err
+		case <-time.After(gracefulShutdownDuration):
+			if err := cmd.Process.Kill(); err != nil {
+				return fmt.Errorf("failed to kill worker: %w", err)
+			}
+			return <-runErrorChan
+		}
 	case err := <-runErrorChan:
 		if err == nil {
 			r.logger.Info("Worker gracefully stopped")
@@ -323,21 +370,21 @@ func (r *Runner) StartWorker(ctx context.Context, options WorkerOptions) error {
 }
 
 type AllInOneOptions struct {
-	RunOptions         RunOptions
-	StartWorkerOptions WorkerOptions
+	WorkerOptions WorkerOptions
 }
 
 // AllInOne run an all-in-one scenario (StartWorker + Run)
 func (r *Runner) AllInOne(ctx context.Context, options AllInOneOptions) error {
 	ctx, cancel := context.WithCancel(ctx)
-	workerErrChan := make(chan error)
-	runnerErrChan := make(chan error)
+	defer cancel()
+	workerErrChan := make(chan error, 1)
+	runnerErrChan := make(chan error, 1)
 
 	go func() {
-		workerErrChan <- r.StartWorker(ctx, options.StartWorkerOptions)
+		workerErrChan <- r.RunWorker(ctx, options.WorkerOptions)
 	}()
 	go func() {
-		runnerErrChan <- r.Run(ctx, options.RunOptions)
+		runnerErrChan <- r.Run(ctx)
 	}()
 
 	var workerErr error
@@ -345,12 +392,10 @@ func (r *Runner) AllInOne(ctx context.Context, options AllInOneOptions) error {
 
 	select {
 	case workerErr = <-workerErrChan:
-		r.logger.Errorf("Worker exited prematurely (error: %v)", workerErr)
-		cancel()
-		runErr = <-runnerErrChan
+		return fmt.Errorf("Worker exited prematurely: %w", workerErr)
 	case runErr = <-runnerErrChan:
 		if runErr != nil {
-			r.logger.Errorf("Run completed with: %v", runErr)
+			r.logger.Errorf("Run completed with: %e", runErr)
 		}
 		cancel()
 		workerErr = <-workerErrChan
@@ -358,9 +403,9 @@ func (r *Runner) AllInOne(ctx context.Context, options AllInOneOptions) error {
 
 	if workerErr != nil {
 		if runErr != nil {
-			return fmt.Errorf("both run and worker completed with errors, worker error: %v, run error: %v", workerErr, runErr)
+			return fmt.Errorf("both run and worker completed with errors, worker error: %v, run error: %w", workerErr, runErr)
 		}
-		return fmt.Errorf("worker finished with error: %v", workerErr)
+		return fmt.Errorf("worker finished with error: %w", workerErr)
 	}
 	return runErr
 }
